rearrange rst files and doc formatting
