--- docs/code_reference/ax_instantiation_utils.rst
@@ -1,8 +1,0 @@
-##############
-Ax Utils
-##############
-
-.. automodule:: boa.ax_instantiation_utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
