+++ docs/code_reference/code_reference.rst
@@ -2,16 +2,49 @@
 Code reference
 ##############
 
-.. toctree::
-    :maxdepth: 1
 
-    ax_instantiation_utils
-    wrapper_utils
-    metaclasses
-    runner
-    storage
-    utils
-    wrapper
-    wrapper_utils
-    metrics
+Wrapping tools
+================
+
+.. autosummary::
+    :toctree: api
+    :template: custom_module_template_short.rst
+
+    boa.wrapper
+    boa.wrapper_utils
+
+
+Metrics
+================
+
+.. autosummary::
+    :toctree: api
+    :template: custom_module_template_short.rst
+
+    boa.metrics.metrics
+    boa.metrics.modular_metric
+    boa.metrics.metric_funcs
+
+
+Saving and Loading your Experiment
+===================================
+
+.. autosummary::
+    :toctree: api
+    :template: custom_module_template_short.rst
+
+    boa.storage
+
+
+Advanced Usage/Direct Python Access
+====================================
+
+.. autosummary::
+    :toctree: api
+    :template: custom_module_template_short.rst
+
+    boa.ax_instantiation_utils
+    boa.runner
+    boa.utils
+    boa.metaclasses
