--- docs/code_reference/metaclasses.rst
@@ -1,8 +1,0 @@
-##############
-Meta-classes
-##############
-
-.. automodule:: boa.metaclasses
-    :members:
-    :undoc-members:
-    :show-inheritance:
