--- docs/code_reference/metrics.rst
@@ -1,7 +1,0 @@
-Metrics
-===========
-
-.. automodule::  boa.metrics.metrics
-   :members:
-   :undoc-members:
-   :show-inheritance:
