--- docs/code_reference/runner.rst
@@ -1,8 +1,0 @@
-##############
-Runner
-##############
-
-.. automodule:: boa.runner
-    :members:
-    :undoc-members:
-    :show-inheritance:
