--- docs/code_reference/storage.rst
@@ -1,8 +1,0 @@
-##############
-Storage
-##############
-
-.. automodule:: boa.storage
-    :members:
-    :undoc-members:
-    :show-inheritance:
