--- docs/code_reference/utils.rst
@@ -1,8 +1,0 @@
-##############
-Utils
-##############
-
-.. automodule:: boa.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
