--- docs/code_reference/wrapper.rst
@@ -1,7 +1,0 @@
-Boa Wrapper
-===========
-
-.. automodule::  boa.wrapper
-   :members:
-   :undoc-members:
-   :show-inheritance:
