--- docs/code_reference/wrapper_utils.rst
@@ -1,10 +1,0 @@
-##############
-Wrapper Utils
-##############
-
-.. automodule:: boa.wrapper_utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
